fix(brute-worker): cast i64 to f64 before D1 bind to avoid BigInt error
wasm-bindgen converts Rust i64 to JS BigInt, which Cloudflare D1 rejects
with D1_TYPE_ERROR. Cast all i64 values to f64 before .into() across all
D1 bind call sites: timestamps, time-series buckets, limit/since params,
checked_at, cutoff_ms, and domain_total.

+++ brute-worker/src/analytics/d1.rs
@@ -90,7 +90,7 @@ impl BruteAnalytics for D1Analytics {
                 "INSERT INTO top_hourly (timestamp, amount) VALUES (?1, 1)
                  ON CONFLICT (timestamp) DO UPDATE SET amount = amount + 1",
             )
-            .bind(&[hour_bucket.into()])
+            .bind(&[(hour_bucket as f64).into()])
             .map_err(|e| BruteError::Database(e.to_string()))?,
 
             // top_daily
@@ -98,7 +98,7 @@ impl BruteAnalytics for D1Analytics {
                 "INSERT INTO top_daily (timestamp, amount) VALUES (?1, 1)
                  ON CONFLICT (timestamp) DO UPDATE SET amount = amount + 1",
             )
-            .bind(&[day_bucket.into()])
+            .bind(&[(day_bucket as f64).into()])
             .map_err(|e| BruteError::Database(e.to_string()))?,
         ];
